requirements.txt
Django==6.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
pandas==3.0.0
openpyxl==3.1.5
django-extensions==4.1